+++ spikeinterface/core/testing_tools.py
