+++ spikeinterface/core/tests/test_job_tools.py
@@ -1,6 +1,6 @@
 
 
-from spikeinterface.core.tests.testing_tools import generate_recording
+from spikeinterface.core.testing_tools import generate_recording
 
 from spikeinterface.core.job_tools import divide_segment_into_chunks, ensure_n_jobs, ensure_chunk_size, ChunkRecordingExecutor
