+++ spikeinterface/toolkit/tests/test_utils.py
@@ -4,7 +4,7 @@ import pytest
 
 import numpy as np
 
-from spikeinterface.core.tests.testing_tools import generate_recording
+from spikeinterface.core.testing_tools import generate_recording
 
 
 from spikeinterface.toolkit.utils import (get_random_data_chunks,
